+++ src/elements/e-vas-input.vue
@@ -296,7 +296,7 @@
       background-color: var(--vas-theme-background-content);
       color: var(--vas-theme-text-color);
       font-family: variables.$vas-font-family--primary;
-      font-size: variables.$vas-font-size--label;
+      font-size: var(--vas-font-size-label);
 
       // disable iPhone styling
       -webkit-appearance: none;
@@ -350,7 +350,7 @@
       display: flex;
       transform: translateY(-50%);
       color: var(--vas-theme-text-color-muted);
-      font-size: variables.$vas-font-size--base;
+      font-size: var(--vas-font-size-base);
     }
 
     &__slot-wrapper {
@@ -365,7 +365,7 @@
     &__slot {
       display: flex;
       color: var(--vas-theme-text-color-muted);
-      font-size: variables.$vas-font-size--base;
+      font-size: var(--vas-font-size-base);
     }
 
     // active